Added custom build of QGroundControl
QGroundControl is a software made by Dronecode that allows users to connect and control drones that support the MAVLink protocol.

QGroundControl has most of the things we need to show to the AUVSI judges during the competition. However, we need to implement a few more visuals based on the data provided by the interop
server. Since QGroundControl is open-source, we are able to implement the remaining visual elements through our own custom plugins.