Initial ML setup
Implemented a CNN for the purpose of identifying a AUVSI SUAS object's color using an architecture published in a report online. Implementation uses PyTorch as the ML framework. Implementation includes the model, a Jupyter Notebook for training, and a example file to run model inference on Intel Neural Compute Stick 2.